Move test to cli ### What's done: * Move test to cli
